homework61

+++ lecture014-java06/Homework/Homework6/src/Product.java
@@ -3,11 +3,11 @@ public class Product {
     public String name;
     public int quantity;
     public ProductUnit unit;
-    public double purchasePrice;
-    public double sellingPrice;
+    public int purchasePrice;
+    public int sellingPrice;
     public int sellingQty;
 
-    public Product(int id, String name, int quantity, ProductUnit unit, double purchasePrice, double sellingPrice, int sellingQty) {
+    public Product(int id, String name, int quantity, ProductUnit unit, int purchasePrice, int sellingPrice, int sellingQty) {
         this.id = id;
         this.name = name;
         this.quantity = quantity;
